src/main/kotlin/com/basic/auth/basic/auth/service/user/adapter/security/CustomUserDetailsService.kt
package com.basic.auth.basic.auth.service.user.adapter.security

import com.basic.auth.basic.auth.service.user.port.driven.persistence.UserRepository
import org.springframework.security.core.userdetails.UserDetails
import org.springframework.security.core.userdetails.UserDetailsService
import org.springframework.security.core.userdetails.UsernameNotFoundException
import org.springframework.stereotype.Service

@Service
class CustomUserDetailsService(
    private val userRepository: UserRepository,
) : UserDetailsService {
    override fun loadUserByUsername(username: String): UserDetails {
        val user = userRepository.findByEmail(username)
            ?: throw UsernameNotFoundException("User not found with email: $username")
        return AuthenticatedUser(user)
    }
}
